refactor(FavoriteAcceptanceTest): 즐겨찾기 생성 인수테스트 리팩터링
