+++ README.md
@@ -7,7 +7,7 @@
 - [ ] 권한이 없을 떄 권한이 필요한 API 호출 시 401 코드 응답
   - [ ] 즐겨 찾기 기능 구현
     - [x] 정상적인 시작, 도착 지점을 입력하면 즐겨찾기 생성
-    - [ ] 비정상적인 시작, 도착 지점을 입력하면 예외를 발생
+    - [x] 비정상적인 시작, 도착 지점을 입력하면 예외를 발생
     - [ ] 즐겨찾기 조회 기능
     - [ ] 정상적인 즐겨찾기 번호를 입력하면 삭제 가능
     - [ ] 비정상적인 즐겨찾기 번호를 입력하면 예외를 발생
