+++ src/test/java/nextstep/subway/acceptance/FavoriteAcceptanceTest.java
@@ -1,25 +1,22 @@
 package nextstep.subway.acceptance;
 
-import io.restassured.RestAssured;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
 import nextstep.subway.exception.NotFoundStationException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 import java.util.HashMap;
 import java.util.Map;
 
+import static nextstep.subway.acceptance.FavoriteSteps.즐겨찾기_생성_성공;
+import static nextstep.subway.acceptance.FavoriteSteps.즐겨찾기_생성_실패_확인;
+import static nextstep.subway.acceptance.FavoriteSteps.즐겨찾기_생성_요청;
 import static nextstep.subway.acceptance.LineSteps.지하철_노선_생성_요청;
 import static nextstep.subway.acceptance.LineSteps.지하철_노선에_지하철_구간_생성_요청;
 import static nextstep.subway.acceptance.MemberSteps.관리자_로그인_되어_있음;
 import static nextstep.subway.acceptance.MemberSteps.유저_로그인_되어_있음;
 import static nextstep.subway.acceptance.StationSteps.지하철역_생성_요청;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
 
 public class FavoriteAcceptanceTest extends AcceptanceTest {
 
@@ -61,63 +58,28 @@ public class FavoriteAcceptanceTest extends AcceptanceTest {
         지하철_노선에_지하철_구간_생성_요청(accessToken, 신분당선, createSectionCreateParams(강남역, 양재역, 3));
         지하철_노선에_지하철_구간_생성_요청(accessToken, 삼호선, createSectionCreateParams(교대역, 남부터미널역, 5));
         지하철_노선에_지하철_구간_생성_요청(accessToken, 삼호선, createSectionCreateParams(남부터미널역, 양재역, 5));
+
+        accessToken = 유저_로그인_되어_있음();
     }
 
     @DisplayName("즐겨찾기 생성에 성공한다")
     @Test
     public void create_favorite_success() {
-        // given
-        Map<String, String> params = new HashMap<>();
-        params.put("source", 교대역 + "");
-        params.put("target", 강남역 + "");
-
-        accessToken = 유저_로그인_되어_있음();
-
         // when
-        ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .auth().oauth2(accessToken)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(params)
-                .when()
-                .post("/favorites")
-                .then().log().all()
-                .extract();
+        var response = 즐겨찾기_생성_요청(accessToken, 교대역, 강남역);
 
         // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+        즐겨찾기_생성_성공(response);
     }
 
     @DisplayName("비정상적인 시작, 도착 지점을 입력하면 예외를 발생시킨다")
     @Test
     public void create_favorite_fail_by_wrong_station() {
-        // given
-        Map<String, String> params = new HashMap<>();
-        params.put("source", 없는역 + "");
-        params.put("target", 강남역 + "");
-
-        accessToken = 유저_로그인_되어_있음();
-
         // when
-        ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .auth().oauth2(accessToken)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(params)
-                .when()
-                .post("/favorites")
-                .then().log().all()
-                .extract();
+        var response = 즐겨찾기_생성_요청(accessToken, 없는역, 강남역);
 
         // then
         즐겨찾기_생성_실패_확인(response, HttpStatus.NOT_FOUND, NotFoundStationException.MESSAGE);
-    }
-
-    public static void 즐겨찾기_생성_실패_확인(ExtractableResponse<Response> response, HttpStatus status, String message) {
-        assertAll(
-                () -> assertThat(response.jsonPath().getString("message")).isEqualTo(message),
-                () -> assertThat(response.jsonPath().getInt("status")).isEqualTo(status.value())
-        );
     }
 
     private Map<String, String> createSectionCreateParams(Long upStationId, Long downStationId, Integer distance) {
