+++ src/test/java/nextstep/subway/acceptance/FavoriteSteps.java
@@ -1,0 +1,43 @@
+package nextstep.subway.acceptance;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+public class FavoriteSteps {
+
+    public static ExtractableResponse<Response> 즐겨찾기_생성_요청(String accessToken, Long source, Long target) {
+        Map<String, String> params = new HashMap<>();
+        params.put("source", source + "");
+        params.put("target", target + "");
+
+        return RestAssured
+                .given().log().all()
+                    .auth().oauth2(accessToken)
+                    .contentType(MediaType.APPLICATION_JSON_VALUE)
+                    .body(params)
+                .when()
+                    .post("/favorites")
+                .then().log().all()
+                .extract();
+    }
+
+    public static void 즐겨찾기_생성_성공(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+    }
+
+    public static void 즐겨찾기_생성_실패_확인(ExtractableResponse<Response> response, HttpStatus status, String message) {
+        assertAll(
+                () -> assertThat(response.jsonPath().getString("message")).isEqualTo(message),
+                () -> assertThat(response.jsonPath().getInt("status")).isEqualTo(status.value())
+        );
+    }
+}
